Create README.md
As part of CS567 coursework at NAU, a testing project was undertaken to evaluate a C++ library management system using DeepState. This involved Mutation Analysis and Code Coverage reports. The system's features, such as displaying available books and study rooms, book check-in and return, and room reservation and cancellation, were thoroughly tested. Leveraging DeepState, a comprehensive set of test cases was executed to identify weaknesses and improve system reliability. This project aimed to enhance the effectiveness of the library management system and contribute insights to software testing methodologies in academia and beyond.